expose models.Puzzle._coerce_val as utils.coerce
